+++ operators/dual_view.py
@@ -1,4 +1,4 @@
-import bpy, os
+import bpy
 
 from pathlib import Path
 
@@ -11,10 +11,12 @@ class TEXT_OT_dual_view(bpy.types.Operator):
     @classmethod
     def poll(cls, context):
         space = bpy.context.space_data
-        filepath = bpy.context.area.spaces.active.text.filepath
-        if filepath.strip() == "": return False
-        return ((space.type == 'TEXT_EDITOR')
-                and Path(filepath).suffix == ".fountain")
+        try: 
+            filepath = bpy.context.area.spaces.active.text.filepath
+            if filepath.strip() == "": return False
+            return ((space.type == 'TEXT_EDITOR')
+                    and Path(filepath).suffix == ".fountain")
+        except AttributeError: return False
 
     original_area = None
 
@@ -120,44 +122,3 @@ class TEXT_OT_dual_view(bpy.types.Operator):
 
 
 handler = None
-
-
-def get_space(context):
-    for area in context.screen.areas:
-        if area.type == "TEXT_EDITOR":
-            return area.spaces.active
-
-
-def text_handler(spc, context):
-
-    scene = bpy.context.scene
-    text = bpy.context.area.spaces.active.text
-    line = text.current_line.body
-    current_text = os.path.basename(bpy.context.space_data.text.filepath)
-    if current_text.strip() == "": return
-    current_character = bpy.data.texts[current_text].current_character
-
-    if not text:
-        return
-
-    if scene.last_line is None and scene.last_line_index != text.current_line_index:
-        scene.last_line = line
-        scene.last_line_index = text.current_line_index
-
-    if scene.last_character is None:  # scene.last_character != current_character:
-        scene.last_character = current_character
-
-    if line != scene.last_line or len(line) != len(scene.last_line):
-        bpy.ops.scene.preview_fountain()
-    elif current_character != scene.last_character:
-        bpy.ops.scene.preview_fountain()
-
-    scene.last_line = line
-    scene.last_character = current_character
-
-
-def redraw(context):
-    for window in context.window_manager.windows:
-        for area in window.screen.areas:
-            if area.type == 'TEXT_EDITOR':
-                area.tag_redraw()
