+++ operators/preview_fountain.py
@@ -11,10 +11,12 @@ class SCREENWRITER_OT_preview_fountain(bpy.types.Operator):
     @classmethod
     def poll(cls, context):
         space = bpy.context.space_data
-        filepath = bpy.context.area.spaces.active.text.filepath
-        if filepath.strip() == "": return False
-        return ((space.type == 'TEXT_EDITOR')
-                and Path(filepath).suffix == ".fountain")
+        try: 
+            filepath = bpy.context.area.spaces.active.text.filepath
+            if filepath.strip() == "": return False
+            return ((space.type == 'TEXT_EDITOR')
+                    and Path(filepath).suffix == ".fountain")
+        except AttributeError: return False
 
     def execute(self, context):
         space = bpy.context.space_data
