+++ operators/scene_to_strip.py
@@ -13,9 +13,12 @@ class TEXT_OT_scenes_to_strips(bpy.types.Operator):
     @classmethod
     def poll(cls, context):
         space = bpy.context.space_data
-        filepath = bpy.context.area.spaces.active.text.filepath
-        return ((space.type == 'TEXT_EDITOR') and
-                Path(filepath).suffix == ".fountain")
+        try: 
+            filepath = bpy.context.area.spaces.active.text.filepath
+            if filepath.strip() == "": return False
+            return ((space.type == 'TEXT_EDITOR')
+                    and Path(filepath).suffix == ".fountain")
+        except AttributeError: return False
 
     def execute(self, context):
